pyproject.toml
[project]
name = "lattica-core"
version = "0.1.0"
description = " Lattica core contracts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "eth_account>=0.13",
  "pyyaml>=6.0",
  "titanoboa>=0.1.9",
  "vyper>=0.4.3"
]

[dependency-groups]
dev = ["mamushi", "pytest", "types-PyYAML", "vyper-lsp"]

[tool.mamushi]
line_length = 100

[tool.pytest]
addopts = ["--tb=short", "-v"]
python_classes = ["Test*"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
testpaths = ["tests"]

[tool.uv.sources]
vyper-lsp = { git = "https://github.com/vyperlang/vyper-lsp.git", rev = "43116b3ca8e5a1f5a1c08adf448fd2d09daaf4dc" }
